Update master branch with 1.20.1-dev changes

+++ src/main/java/tech/anonymoushacker1279/immersiveweapons/blockentity/DamageableBlockEntity.java
@@ -128,7 +128,7 @@ public class DamageableBlockEntity extends BlockEntity {
 		currentStage = state.getValue(damageStage);
 		if (repairStack.getItem() == repairItem && currentStage > 0 && currentStage <= stages) {
 			if (health < maxHealth) {
-				health += maxHealth / (double) (stages + 1);
+				health += (int) (maxHealth / (double) (stages + 1));
 
 				level.setBlockAndUpdate(pos, state.setValue(damageStage, currentStage - 1));
